Cleaned up a few files, updated boilerplate to latest version 0.18.0
MUST UPDATE DEPENDENCIES TO WORK
