+++ .eslintrc.js
@@ -1,0 +1,10 @@
+module.exports = {
+    extends: 'erb',
+    settings: {
+      'import/resolver': {
+        webpack: {
+          config: require.resolve('./configs/webpack.config.eslint.js')
+        }
+      }
+    }
+  }
