+++ app/app.html
@@ -1,9 +1,8 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <meta charset="utf-8">
-    <title>TestApp</title>
-    <link rel="stylesheet" href="../node_modules/antd/dist/antd.min.css">
+    <meta charset="utf-8" />
+    <link rel="stylesheet" href="../node_modules/antd/dist/antd.min.css" />
     <script>
       (function() {
         if (!process.env.HOT) {
@@ -13,13 +12,11 @@
           // HACK: Writing the script path should be done with webpack
           document.getElementsByTagName('head')[0].appendChild(link);
         }
-      }());
+      })();
     </script>
   </head>
   <body>
-    <div id="root">
-    
-    </div>
+    <div id="root"></div>
     <script>
       {
         const scripts = [];
@@ -33,7 +30,7 @@
         // Dynamically insert the bundled app script in the renderer process
         const port = process.env.PORT || 1212;
         scripts.push(
-          (process.env.HOT)
+          process.env.HOT
             ? 'http://localhost:' + port + '/dist/renderer.dev.js'
             : './dist/renderer.prod.js'
         );
