+++ app/components/ExportData.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import ReactToPdf from "react-to-pdf";
-import NivoGraph from '../components/NivoSampleGraph';
+import NivoGraph from './GraphData';
 
 type Props = {};
 
@@ -11,22 +11,24 @@ export default class ExportData extends Component {
 
 
   render() {
-    
+
     const ref = React.createRef();
     const options = {
       orientation: 'landscape'
-  };
+    };
 
     return (
       <div>
         <ReactToPdf targetRef={ref} filename="div-blue.pdf" options={options} x={.5} y={.5}>
-        {({toPdf}) => (
+          {({ toPdf }) => (
             <button onClick={toPdf}>Generate pdf</button>
-        )}
+          )}
         </ReactToPdf>
 
-      <div style={{width: 1250, height: 1250}} ref={ref}><NivoGraph csvData={this.props.csvData} /></div>
-    </div>
+        <div style={{ width: 1250, height: 1250 }} ref={ref}>
+          <NivoGraph csvData={this.props.csvData} />
+        </div>
+      </div>
     );
   }
