+++ app/constants/routes.json
@@ -1,7 +1,3 @@
 {
-  "HOME": "/",
-  "GRAPH": "/counter",
-  "TABLE": "/table",
-  "EXPORT": "/export",
-  "SETTINGS": "/setting"
+  "HOME": "/"
 }
