+++ app/containers/App.js
@@ -10,6 +10,6 @@ export default class App extends React.Component<Props> {
 
   render() {
     const { children } = this.props;
-    return <React.Fragment>{children}</React.Fragment>;
+    return <>{children}</>;
   }
 }
