+++ app/containers/HomePage.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import routes from '../constants/routes';
 import FileUploadForm from '../components/FileUploadForm';
-import NivoGraph from '../components/NivoSampleGraph';
+import NivoGraph from '../components/GraphData';
 import TableData from '../components/TableData';
 import ExportData from '../components/ExportData';
 //import SideBar from 'react-fixed-sidebar';
@@ -34,10 +34,10 @@ export default class HomePage extends Component<Props> {
   onCollapse = collapsed => {
     console.log(collapsed);
     this.setState({ collapsed });
-    
+
   };
 
- 
+
   //Handles button state and gets csv data
   handleCsvData = csvData => {
     //console.log(csvData);
@@ -83,7 +83,7 @@ export default class HomePage extends Component<Props> {
   handleTable = showTable => {
     console.log(showTable);
     this.setState(
-      { 
+      {
         showGraph: false,
         title: "Table Page",
         subTitle: "",
@@ -92,23 +92,23 @@ export default class HomePage extends Component<Props> {
         showImport: false,
         showExport: false
       }
-     );
+    );
   }
-//Handles table generation state
-handleExport = showExport => {
-  console.log(showExport);
-  this.setState(
-    { 
-      showGraph: false,
-      showTable: false,
-      title: "Export Page",
-      subTitle: "",
-      showExport,
-      key: "4",
-      showImport: false
-    }
-   );
-}
+  //Handles table generation state
+  handleExport = showExport => {
+    console.log(showExport);
+    this.setState(
+      {
+        showGraph: false,
+        showTable: false,
+        title: "Export Page",
+        subTitle: "",
+        showExport,
+        key: "4",
+        showImport: false
+      }
+    );
+  }
   clearData = csvData => {
     console.log("Clearing CSV data");
     this.setState(
@@ -132,66 +132,67 @@ handleExport = showExport => {
 
     return (
       <Layout style={{ minHeight: '100vh' }}>
-        
-        <Sider collapsible collapsed={this.state.collapsed} onCollapse={this.onCollapse}>
-        <Sider style={{
-          overflow: 'auto',
-          height: '100vh',
-          position: 'fixed',
-          left: 0,}}>
 
-          <div className="logo" >
-          <Menu theme="dark" defaultSelectedKeys={['1']} mode="inline" selectedKeys={this.state.key}>
-            <Menu.Item key="1">
-              <span><Button type="link" icon="home" onClick={this.handleHome}>Home</Button></span>
-            </Menu.Item>
-            <Menu.Item key="2">
-              <span><Button type="link" icon="bar-chart" onClick={this.handleGraph} disabled={!this.state.button}>Graphs</Button></span>
-            </Menu.Item>
-            <Menu.Item key="3">
-              <span><Button type="link" icon="table" onClick={this.handleTable} disabled={!this.state.button}>Tables</Button></span>
-            </Menu.Item>
-            <Menu.Item key="4">
-              <span><Button type="link" icon="export" onClick={this.handleExport} disabled={!this.state.button}>Export</Button></span>
-            </Menu.Item>
-            <Menu.Item key="5">
-              <span><Button type="link" icon="setting" disabled={!this.state.button}>Settings</Button></span>
-            </Menu.Item>
-          </Menu>
-          </div>
-        </Sider>
+        <Sider collapsible collapsed={this.state.collapsed} onCollapse={this.onCollapse}>
+          <Sider style={{
+            overflow: 'auto',
+            height: '100vh',
+            position: 'fixed',
+            left: 0,
+          }}>
+
+            <div className="logo" >
+              <Menu theme="dark" defaultSelectedKeys={['1']} mode="inline" selectedKeys={this.state.key}>
+                <Menu.Item key="1">
+                  <span><Button type="link" icon="home" onClick={this.handleHome}>Home</Button></span>
+                </Menu.Item>
+                <Menu.Item key="2">
+                  <span><Button type="link" icon="bar-chart" onClick={this.handleGraph} disabled={!this.state.button}>Graphs</Button></span>
+                </Menu.Item>
+                <Menu.Item key="3">
+                  <span><Button type="link" icon="table" onClick={this.handleTable} disabled={!this.state.button}>Tables</Button></span>
+                </Menu.Item>
+                <Menu.Item key="4">
+                  <span><Button type="link" icon="export" onClick={this.handleExport} disabled={!this.state.button}>Export</Button></span>
+                </Menu.Item>
+                <Menu.Item key="5">
+                  <span><Button type="link" icon="setting" disabled={!this.state.button}>Settings</Button></span>
+                </Menu.Item>
+              </Menu>
+            </div>
+          </Sider>
         </Sider>
         <Layout>
           <Header style={{ background: '#fff', padding: 0 }}>
             <PageHeader title={this.state.title} subTitle={this.state.subTitle} />
           </Header>
-          
+
           <Content style={{ margin: '0 16px' }}>
-          
-            <div style={{  display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center" }}>
+
+            <div style={{ display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center" }}>
               {this.state.showImport ?
-                
-                <div style={{ padding: 24, display: "flex", flexDirection: "column",background: '#fff',  alignItems:"center" }}> 
-                    
-                    <img src="Dalhousie-Header4.png" alt="DMU"/>
-                    <div style={{ padding: 10, display: "flex",flexDirection: "column",background: '#1890ff',  alignItems:"center", borderRadius:5, margin: 20}}>
-                      <p>WELCOME!</p>
-                      <h3>Welcome to the Dalhousie Youth Services App!<br></br>This application generates modifiable reports that <br></br>
+
+                <div style={{ padding: 24, display: "flex", flexDirection: "column", background: '#fff', alignItems: "center" }}>
+
+                  <img src="img/Dalhousie-Header4.png" alt="DMU" />
+                  <div style={{ padding: 10, display: "flex", flexDirection: "column", background: '#1890ff', alignItems: "center", borderRadius: 5, margin: 20 }}>
+                    <p>WELCOME!</p>
+                    <h3>Welcome to the Dalhousie Youth Services App!<br></br>This application generates modifiable reports that <br></br>
                       contain graphs and chart that show monthly, quaterly, <br></br>
                       and/or yearly reports. Please put your desired <br></br>CSV file for converting
                      </h3>
-                    
-                  <FileUploadForm getCsvData={this.handleCsvData} />
-                  <Button onClick={this.clearData}>Clear CSV</Button>
+
+                    <FileUploadForm getCsvData={this.handleCsvData} />
+                    <Button onClick={this.clearData}>Clear CSV</Button>
                   </div>
-              </div>: null}
-             </div>
+                </div> : null}
+            </div>
             {this.state.showGraph ? <div style={{ padding: 24 }}> <NivoGraph csvData={this.state.csvData} /> </div> : null}
             {this.state.showTable ? <TableData csvData={this.state.csvData} /> : null}
             {this.state.showExport ? <ExportData csvData={this.state.csvData} /> : null}
           </Content>
           <Footer style={{ textAlign: 'center' }}>DMUⒸ REACT VERSION: {REACT_VERSION}</Footer>
-          </Layout>
+        </Layout>
       </Layout>
     );
   }
