--- app/containers/ImportPage.js
@@ -1,39 +1,0 @@
-// @flow
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-import routes from '../constants/routes';
-import APP_LAYOUT from '../components/AppLayout';
-import FileUploadForm from '../components/FileUploadForm';
-import { PageHeader, Layout, Menu, Breadcrumb, Icon } from 'antd';
-
-type Props = {};
-
-const { Header, Content, Footer, Sider } = Layout;
-const { SubMenu } = Menu;
-
-export default class HomePage extends Component<Props> {
-  props: Props;
-
-  state = {
-    collapsed: false,
-  };
-
-  //Handles the responsive sidebar
-  onCollapse = collapsed => {
-    console.log(collapsed);
-    this.setState({ collapsed });
-  };
-
-  /*
-    Docs for Antd Layouts: https://ant.design/components/layout/
-
-    Summary:  Renders the app layout, including the sidebar. Also loads the FileUploadForm component
-              into the a div displayed on the main page.
-  */
-
-  render() {
-    return (
-      
-    );
-  }
-}
