+++ app/containers/Root.js
@@ -1,7 +1,8 @@
 // @flow
-import React, { Component } from 'react';
+import React from 'react';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'connected-react-router';
+import { hot } from 'react-hot-loader/root';
 import type { Store } from '../reducers/types';
 import Routes from '../Routes';
 
@@ -10,15 +11,12 @@ type Props = {
   history: {}
 };
 
-export default class Root extends Component<Props> {
-  render() {
-    const { store, history } = this.props;
-    return (
-      <Provider store={store}>
-        <ConnectedRouter history={history}>
-          <Routes />
-        </ConnectedRouter>
-      </Provider>
-    );
-  }
-}
+const Root = ({ store, history }: Props) => (
+  <Provider store={store}>
+    <ConnectedRouter history={history}>
+      <Routes />
+    </ConnectedRouter>
+  </Provider>
+);
+
+export default hot(Root);
