--- app/containers/SideBarPage.js
@@ -1,13 +1,0 @@
-// @flow
-import React, { Component } from 'react';
-import SideBar from '../components/SideBar';
-
-type Props = {};
-
-export default class SideBarPage extends Component<Props> {
-    props: Props;
-
-    render() {
-        return <SideBar />;
-    }
-}
