+++ app/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "electron-react-boilerplate",
+    "productName": "electron-react-boilerplate",
+    "version": "0.18.0",
+    "description": "Electron application boilerplate based on React, React Router, Webpack, React Hot Loader for rapid application development",
+    "main": "./main.prod.js",
+    "author": {
+      "name": "Electron React Boilerplate Maintainers",
+      "email": "electronreactboilerplate@gmail.com",
+      "url": "https://github.com/electron-react-boilerplate"
+    },
+    "scripts": {
+      "electron-rebuild": "node -r ../internals/scripts/BabelRegister.js ../internals/scripts/ElectronRebuild.js",
+      "postinstall": "yarn electron-rebuild"
+    },
+    "license": "MIT",
+    "dependencies": {}
+  }
