+++ app/reducers/index.js
@@ -1,10 +1,11 @@
 // @flow
 import { combineReducers } from 'redux';
 import { connectRouter } from 'connected-react-router';
+import type { HashHistory } from 'history';
 import counter from './counter';
 
-export default function createRootReducer(history: History) {
-  return combineReducers({
+export default function createRootReducer(history: HashHistory) {
+  return combineReducers<{}, *>({
     router: connectRouter(history),
     counter
   });
