+++ app/store/configureStore.prod.js
@@ -12,7 +12,11 @@ const router = routerMiddleware(history);
 const enhancer = applyMiddleware(thunk, router);
 
 function configureStore(initialState?: counterStateType) {
-  return createStore(rootReducer, initialState, enhancer);
+  return createStore<*, counterStateType, *>(
+    rootReducer,
+    initialState,
+    enhancer
+  );
 }
 
 export default { configureStore, history };
