+++ configs/webpack.config.base.js
@@ -4,10 +4,10 @@
 
 import path from 'path';
 import webpack from 'webpack';
-import { dependencies } from '../package.json';
+import { dependencies as externals } from '../app/package.json';
 
 export default {
-  externals: [...Object.keys(dependencies || {})],
+  externals: [...Object.keys(externals || {})],
 
   module: {
     rules: [
@@ -34,7 +34,8 @@ export default {
    * Determine the array of extensions that should be used to resolve modules.
    */
   resolve: {
-    extensions: ['.js', '.jsx', '.json']
+    extensions: ['.js', '.jsx', '.json'],
+    modules: [path.join(__dirname, '..', 'app'), 'node_modules']
   },
 
   plugins: [
