+++ internals/scripts/BabelRegister.js
@@ -1,0 +1,5 @@
+const path = require('path');
+
+require('@babel/register')({
+  cwd: path.join(__dirname, '..', '..')
+});
