+++ renovate.json
@@ -1,9 +1,3 @@
 {
-  "extends": ["config:base"],
-  "rangeStrategy": "bump",
-  "baseBranches": ["next"],
-  "automerge": true,
-  "major": {
-    "automerge": false
-  }
+  "extends": ["bliss"]
 }
